refactor(users): extract member-import parser into lib/server/member-import.ts
users-service.ts (~872 lines) mixed user CRUD with ~420 lines of CSV/ODT/XLSX
parsing and normalization. Move all parse/normalize helpers, their types and
constants, and the exceljs/fflate imports into a new lib/server/member-import.ts.
users-service.ts now only orchestrates the DB import (importMembersFromCsv /
importMembersFromSource / importMembersFromNormalizedRows) and imports the
parsing entry points from the new module. Pure move — no behavior change.

users-service.ts: 872 -> 449 lines.

Refs KIM-407
